+++ cmd/root.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"log"
 	"os"
+	"strings"
 
 	"github.com/spf13/cobra"
 
@@ -15,9 +16,10 @@ import (
 	"github.com/giantswarm/backstage-catalog-importer/pkg/input/helmchart"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/input/repositories"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/input/teams"
+	bscatalog "github.com/giantswarm/backstage-catalog-importer/pkg/output/bscatalog/v1alpha1"
+	"github.com/giantswarm/backstage-catalog-importer/pkg/output/catalog/component"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/output/catalog/group"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/output/export"
-	"github.com/giantswarm/backstage-catalog-importer/pkg/output/legacy"
 )
 
 var rootCmd = &cobra.Command{
@@ -39,6 +41,9 @@ const (
 
 func init() {
 	rootCmd.PersistentFlags().StringP("output", "o", ".", "Output directory path")
+	rootCmd.Flags().StringP("chart-repo-prefix", "", "charts/giantswarm", "Prefix for chart repositories in the OCI registries")
+	rootCmd.Flags().StringP("public-oci-registry", "", "gsoci.azurecr.io", "Host name of the public OCI registry")
+	rootCmd.Flags().StringP("private-oci-registry", "", "gsociprivate.azurecr.io", "Host name of the private OCI registry")
 
 	rootCmd.AddCommand(appcatalogs.Command)
 	rootCmd.AddCommand(charts.Command)
@@ -55,6 +60,24 @@ func Execute() {
 
 func runRoot(cmd *cobra.Command, args []string) {
 	path, err := cmd.PersistentFlags().GetString("output")
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	repoPrefix, err := cmd.Flags().GetString("chart-repo-prefix")
+	if err != nil {
+		log.Fatal(err)
+	}
+	// Remove leading and trailing slash if present
+	repoPrefix = strings.TrimPrefix(repoPrefix, "/")
+	repoPrefix = strings.TrimSuffix(repoPrefix, "/")
+
+	publicOciRegistry, err := cmd.Flags().GetString("public-oci-registry")
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	privateOciRegistry, err := cmd.Flags().GetString("private-oci-registry")
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -102,14 +125,14 @@ func runRoot(cmd *cobra.Command, args []string) {
 		log.Printf("Processing %d repos of team %q\n", len(list.Repositories), list.OwnerTeamName)
 
 		for _, repo := range list.Repositories {
+
+			ociRegistry := publicOciRegistry
 			isPrivate, err := repoService.GetIsPrivate(repo.Name)
 			if err != nil {
 				log.Fatalf("Error: %v", err)
 			}
-
-			hasCircleCi, err := repoService.GetHasCircleCI(repo.Name)
-			if err != nil {
-				log.Fatalf("Error: %v", err)
+			if isPrivate {
+				ociRegistry = privateOciRegistry
 			}
 
 			hasReadme, err := repoService.GetHasReadme(repo.Name)
@@ -119,6 +142,7 @@ func runRoot(cmd *cobra.Command, args []string) {
 
 			// Fetch Helm chart info if available.
 			var charts []*helmchart.Chart
+			var hasDeployableChart bool
 			{
 				numCharts, err := repoService.GetNumHelmCharts(repo.Name)
 				if err != nil {
@@ -139,6 +163,9 @@ func runRoot(cmd *cobra.Command, args []string) {
 								log.Printf("WARN - %s - error parsing helm chart %s: %v", repo.Name, chartName, err)
 							} else {
 								charts = append(charts, chart)
+								if chart.Type == "application" || chart.Type == "" {
+									hasDeployableChart = true
+								}
 							}
 						}
 					}
@@ -169,22 +196,80 @@ func runRoot(cmd *cobra.Command, args []string) {
 				}
 			}
 
-			ent := legacy.CreateComponentEntity( //nolint:staticcheck
-				repo,
-				list.OwnerTeamName,
-				repoService.MustGetDescription(repo.Name),
-				repo.System,
-				isPrivate,
-				hasCircleCi,
-				hasReadme,
-				repoService.MustGetDefaultBranch(repo.Name),
-				latestReleaseTime,
-				latestReleaseTag,
-				charts,
-				deps)
+			// Prepare deployment names (default to <name> and <name>-app if not set).
+			deploymentNames := repo.DeploymentNames
+			if len(deploymentNames) == 0 {
+				name := strings.TrimSuffix(repo.Name, "-app")
+				nameWithAppSuffix := fmt.Sprintf("%s-app", name)
+				deploymentNames = []string{
+					name,
+					nameWithAppSuffix,
+				}
+			}
+
+			description := repoService.MustGetDescription(repo.Name)
+			defaultBranch := repoService.MustGetDefaultBranch(repo.Name)
+
+			genLanguage := ""
+			if repo.Gen.Language != "" && repo.Gen.Language != repositories.RepoLanguageGeneric {
+				genLanguage = string(repo.Gen.Language)
+			}
+
+			genFlavors := make([]string, len(repo.Gen.Flavors))
+			for i, flavor := range repo.Gen.Flavors {
+				genFlavors[i] = string(flavor)
+			}
+
+			c, err := component.New(
+				repo.Name,
+				component.WithCircleCiSlug(fmt.Sprintf("github/%s/%s", githubOrganization, repo.Name)),
+				component.WithDefaultBranch(defaultBranch),
+				component.WithDependsOn(deps...),
+				component.WithDeploymentNames(deploymentNames...),
+				component.WithDescription(description),
+				component.WithFlavors(genFlavors...),
+				component.WithGithubProjectSlug(fmt.Sprintf("%s/%s", githubOrganization, repo.Name)),
+				component.WithGithubTeamSlug(list.OwnerTeamName),
+				component.WithHasReadme(hasReadme),
+				component.WithHasReleases(latestReleaseTag != ""),
+				component.WithHelmCharts(charts...),
+				component.WithLanguage(genLanguage),
+				component.WithLatestReleaseTag(latestReleaseTag),
+				component.WithLatestReleaseTime(latestReleaseTime),
+				component.WithLifecycle(string(repo.Lifecycle)),
+				component.WithOwner(list.OwnerTeamName),
+				component.WithPrivate(isPrivate),
+				component.WithSystem(repo.System),
+				component.WithType(repo.ComponentType),
+				component.WithOciRegistry(ociRegistry),
+				component.WithOciRepositoryPrefix(repoPrefix),
+			)
+			if err != nil {
+				log.Fatalf("Could not create component: %s", err)
+			}
+
+			if hasDeployableChart {
+				c.AddTag("helmchart-deployable")
+			}
+
+			// Grafana dashboard link for services.
+			if repo.ComponentType == "service" {
+				urlParts := []string{}
+				for _, d := range deploymentNames {
+					urlParts = append(urlParts, fmt.Sprintf("var-app=%s", d))
+				}
+				c.AddLink(bscatalog.EntityLink{
+					URL:   fmt.Sprintf("https://giantswarm.grafana.net/d/eb617ba1-209a-4d57-9963-1af9a8ddc8d4/general-service-metrics?orgId=1&%s&from=now-24h&to=now", strings.Join(urlParts, "&")),
+					Title: "General service metrics dashboard",
+					Icon:  "dashboard",
+					Type:  "grafana-dashboard",
+				})
+			}
+
+			entity := c.ToEntity()
 			numComponents++
 
-			err = componentExporter.AddEntity(&ent)
+			err = componentExporter.AddEntity(entity)
 			if err != nil {
 				log.Fatalf("Error: %v", err)
 			}