+++ pkg/input/repositories/repositories.go
@@ -14,7 +14,7 @@ import (
 	"time"
 
 	"github.com/giantswarm/microerror"
-	"github.com/google/go-github/v79/github"
+	"github.com/google/go-github/v81/github"
 	"go.yaml.in/yaml/v3"
 	"golang.org/x/oauth2"
 )