+++ pkg/input/teams/teams.go
@@ -5,7 +5,7 @@ import (
 	"context"
 
 	"github.com/giantswarm/microerror"
-	"github.com/google/go-github/v79/github"
+	"github.com/google/go-github/v81/github"
 	"golang.org/x/oauth2"
 )